fix: baitWell source (#169)
* fix: baitWell source

The baitWell source was showing up as undefined because it was looking for Temperature Instance but it is Instance in pgns.h.

Closes #168

* fix: temperature tests for liveWell and baitWell

The instance is ignored for all temperatures except the liveWell and
baitWell sources. For these two sources the instance becomes a part
of the path in SK. The json received from canboatjs contains "Instance"
and not "Temperature Instance". The tests for these two sources are
updated to reflect this.

Closes #169