feature: adds support for alert PGNs 126983 and 126985 (#173)
NMEA has standardized PGNs to handle alerts. The main alert
is handled in 126983 with the alert text handled in 126985.

The alert is sent ~1/sec with the alert text sent ~1/10 sec
or possibly longer. Because the notifications schema only allows
for a single value, we wait until we have data from both PGNs
before we process the update and create the notification.