feature: adds support for user defined temperatures (#179)
* feature: adds support for user defined temperatures

This adds support for those temperature sources not specifically defined by NMEA.
NMEA allows for a user defined temperature source ranging from 129-252.

The new path will be generic.temperatures.userDefined<###>.<instance>.temperature

* test: add flag for data not in SK Specification

Typically tests will validate the test data against the SK
Specification. However if the data does not exist in the spec
it will cause the test to fail.

Adding a boolean flag notInSpec = true to the test data will
cause this part of the validation to be skipped.